Test.QuickCheck.Instances

Description

Instances are provided for the types in the packages:

array

bytestring

case-insensitive

containers

data-fix

OneTuple

old-time

strict

text

text-short

these

time

unordered-containers

uuid

primitive

vector

Since all of these instances are provided as orphans, I recommend that you do not use this library within another library module, so that you don't impose these instances on down-stream consumers of your code.

For information on writing a test-suite with Cabal see http://www.haskell.org/cabal/users-guide/#test-suites