Test.QuickCheck.Instances

Description

Instances are provided for the types in the packages:

array

bytestring

case-insensitive

containers

old-time

text

time

unordered-containers

Since all of these instances are provided as orphans, I recommend that you do not use this library within another library module, so that you don't impose these instances on down-stream consumers of your code.

For information on writing a test-suite with Cabal see http://www.haskell.org/cabal/users-guide/#test-suites